tests/wpt/web-platform-tests/html/semantics/embedded-content/the-canvas-element/fallback.nested.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by tools/gentest.py. -->
<title>Canvas test: fallback.nested</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/2dcontext/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/2dcontext/resources/canvas-tests.css">
<body class="show_output">

<h1>fallback.nested</h1>
<p class="desc">Fallback content containing another canvas (mostly testing parsers)</p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><canvas><p class="fallback">FAIL (fallback content)</p></canvas><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("Fallback content containing another canvas (mostly testing parsers)");
_addTest(function(canvas, ctx) {

_assertSame(canvas.childNodes.length, 2, "canvas.childNodes.length", "2");


});
</script>
